1.0.3 release
Add support for host EVSE "pausing." That is, it's no longer an error if the host EVSE's pilot goes away. Instead, the pause state will be propagated to the cars and they'll be gracefully shut down. Assuming the Hydra is independently powered, the paused state will be maintained until the host EVSE's pilot comes back, and then everything will be restarted.