Update Hydra_EVSE.ino
Now that we're performing relay and ground tests, we can't just flip the power off during a GFI failure. Depending on when the interrupt occurs, we could immediately perform a relay test and fail because the power is off. That winds up masking the "G" error with an "R" error. So we need to make all the data consistent with a relay transition in the GFI interrupt service routine.